fix(Alert): Renamed to AlertBar
BREAKING CHANGE:

Alert component renamed to AlertBar

```jsx
// Before
<Alert></Alert>

// After
<AlertBar></AlertBar>
