+++ src/components/alert-bar/AlertBar.css
@@ -1,21 +1,21 @@
-.ax-alert {
+.ax-alert-bar {
   padding-right: var(--page-padding-horizontal);
   padding-left: var(--page-padding-horizontal);
   color: var(--color-ui-white-noise);
   font-weight: var(--font-weight-medium);
 }
 
-.ax-alert--small {
+.ax-alert-bar--small {
   padding-top: var(--spacing-grid-size--x2);
   padding-bottom: var(--spacing-grid-size--x2);
 }
 
-.ax-alert--medium {
+.ax-alert-bar--medium {
   padding-top: var(--spacing-grid-size--x3);
   padding-bottom: var(--spacing-grid-size--x3);
 }
 
-.ax-alert--success { background-color: var(--color-ui-success); }
-.ax-alert--warning { background-color: var(--color-ui-warning); }
-.ax-alert--error   { background-color: var(--color-ui-error); }
-.ax-alert--info    { background-color: var(--color-ui-info); }
+.ax-alert-bar--success { background-color: var(--color-ui-success); }
+.ax-alert-bar--warning { background-color: var(--color-ui-warning); }
+.ax-alert-bar--error   { background-color: var(--color-ui-error); }
+.ax-alert-bar--info    { background-color: var(--color-ui-info); }
